Picture of Marcus Aurelius <3
Marcus Aurelius was a Roman emperor from 161 to 180 and a Stoic philosopher. He was the last of the rulers known as the Five Good Emperors, and the last emperor of the Pax Romana, an age of relative peace and stability for the Roman Empire. He served as Roman consul in 140, 145, and 161. 

One of the Founding Father of Stoicism, the philosophy I follow.